% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccineWaningMethodEvaluation.R
\docType{package}
\name{VaccineWaningMethodEvaluation-package}
\alias{VaccineWaningMethodEvaluation}
\alias{VaccineWaningMethodEvaluation-package}
\title{VaccineWaningMethodEvaluation: Evaluation of Methods for Estimating Waning of Vaccine Effectiveness}
\description{
A methods research experiment to evaluate various methods for estimating waning of vaccine effectiveness.
}
\keyword{internal}
