% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medon.R
\docType{package}
\name{Medon-package}
\alias{Medon}
\alias{Medon-package}
\title{Medon: Evaluation of Methods for Estimating Waning of Vaccine Effectiveness}
\description{
A methods research experiment to evaluate various methods for estimating waning of vaccine effectiveness.
}
\keyword{internal}
